The original issue only pertained to unit. Hence only do the fix for unit tests.

+++ src/Altinn.Broker.Application/ConfirmDownload/ConfirmDownloadHandler.cs
@@ -20,6 +20,7 @@ using Hangfire;
 using Microsoft.Extensions.Logging;
 
 using OneOf;
+using Microsoft.Extensions.Hosting;
 
 public class ConfirmDownloadHandler(
     IFileTransferRepository fileTransferRepository,
@@ -28,6 +29,7 @@ public class ConfirmDownloadHandler(
     IResourceRepository resourceRepository,
     IAuthorizationService authorizationService,
     IBackgroundJobClient backgroundJobClient,
+    IHostEnvironment hostEnvironment,
     EventBusMiddleware eventBus,
     ILogger<ConfirmDownloadHandler> logger) : IHandler<ConfirmDownloadRequest, Task>
 {
@@ -90,13 +92,14 @@ public class ConfirmDownloadHandler(
                         FileTransferId = request.FileTransferId,
                         PurgeTrigger = PurgeTrigger.AllConfirmedDownloaded
                     }, null, cancellationToken), DateTime.UtcNow.Add(gracePeriod));
+                    backgroundJobClient.Delete(fileTransfer.HangfireJobId); // Old expiry job
                 }
             }
             return Task.CompletedTask;
         }, logger, cancellationToken);
-        if (shouldConfirmAll && resource!.PurgeFileTransferAfterAllRecipientsConfirmed)
+        if (hostEnvironment.IsDevelopment() && shouldConfirmAll && resource!.PurgeFileTransferAfterAllRecipientsConfirmed)
         {
-            backgroundJobClient.Delete(fileTransfer.HangfireJobId); // Performed outside of transaction to avoid issue with Hangfire distributed lock implementation
+            backgroundJobClient.Delete(fileTransfer.HangfireJobId); // Performed outside of transaction in unit tests to avoid issue with Hangfire distributed lock implementation
         }
         return Task.CompletedTask;
     }
